Remove WIN32 warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@101074 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 1ead86f66391dfda630dcdd5cd1e5c00f3c127f7

+++ Online/Online/GaudiOnline/GaudiOnline/MEPRxSvc.h
@@ -203,8 +203,8 @@ namespace LHCb  {
     int       spaceSize()  const            {  return m_MEPBufSize;   }
     int       sourceID() const              {  return m_sourceID;     }
     int       sourceAddr(u_int32_t i)       {  return m_srcAddr[i];   }
-    int       addIncompleteEvent()          {  return m_incEvt++;     }
-    int       totWrongPartID()              {  return m_totWrongPartID;  }   
+    int64_t   addIncompleteEvent()          {  return m_incEvt++;     }
+    int64_t   totWrongPartID()              {  return m_totWrongPartID;  }   
     RXIT ageRx();
     static bool cmpL0ID(MEPRx *r, u_int32_t id);
     void removePkt(void);
